Fix highlighting of results
On our version (2.3.2), you need to tell ES to do highlighting on fields
even if we didn't search against them. This is necessary because we
preprocess our input data, to have an indexable field, and a pretty
presentable field for results display.

We search against X but want to highlight Y for presentation. Normally
ES will only search against X and highlight X as well.